sgd/requirements/dev.txt
-r common.txt
black
ruff
mypy
tox
pytest-django
pytest-randomly
pytest-cov
pytest-structlog
factory-boy
django-silk
django-debug-toolbar
sphinxcontrib-django
